refactor: useHeader에서 useGradient 훅 분리
