+++ src/hooks/useHeader.ts
@@ -1,33 +1,22 @@
-import { authService } from 'apis/firebaseService';
-import { signOut } from 'firebase/auth';
 import { useEffect, useState } from 'react';
-import { useLocation, useNavigate } from 'react-router-dom';
+import { useNavigate } from 'react-router-dom';
 import { useSetRecoilState } from 'recoil';
-import { resetAmplitude } from 'utils/amplitude';
 import { userInfoState } from '../recoil/atoms';
-import useAuth from './useAuth';
+import { useGradient, useAuth } from 'hooks';
+import { signOut } from 'firebase/auth';
+import { authService } from 'apis/firebaseService';
 import { defaultInfo } from './useUpdateProfile';
-
-// 메인 페이지에서 스크롤이 MAIN_SCROLL_Y 값 이상 되면 헤더의 배경색을 투명에서 하얀색으로 변경
-const MAIN_SCROLL_Y = 480;
+import { resetAmplitude } from 'utils/amplitude';
 
 const useHeader = () => {
-  const [hideGradient, setHideGradient] = useState<boolean>(true);
   const [isLoggedIn, setIsLoggedIn] = useState<boolean>(false);
   const [showDropwdown, setShowDropdown] = useState<boolean>(false);
   const setUserInfo = useSetRecoilState(userInfoState);
-  const localUser = useAuth();
 
   const navigate = useNavigate();
-  const location = useLocation();
 
-  const isMain = location.pathname === '/';
-
-  // 높이에 따라 배경색 그라디언트 표시 여부 결정하는 함수
-  const showHeaderGradientBackground = () => {
-    const { scrollY } = window;
-    scrollY > MAIN_SCROLL_Y ? setHideGradient(false) : setHideGradient(true);
-  };
+  const localUser = useAuth();
+  const { isMain, hideGradient } = useGradient();
 
   // 로그아웃
   const handleLogoutClick = () => {
@@ -61,17 +50,6 @@ const useHeader = () => {
   const closeDropdownMenu = () => {
     setShowDropdown(false);
   };
-
-  // 메인 페이지일 경우 스크롤에 따른 배경 그라이언트 함수 이벤트 적용
-  useEffect(() => {
-    if (isMain) {
-      window.addEventListener('scroll', showHeaderGradientBackground);
-    }
-
-    return () => {
-      window.removeEventListener('scroll', showHeaderGradientBackground);
-    };
-  }, [isMain]);
 
   // 유저 로그인 여부 판단
   useEffect(() => {
